Move test helpers out of action build
