+++ packages/action/src/TasitAction.ts
@@ -4,7 +4,6 @@ import Action from "./action";
 
 import Contract from "./contract";
 import Utils from "./contract/Utils";
-import helpers from "./testHelpers/helpers";
 
 export { Contract };
 
@@ -13,7 +12,6 @@ export const TasitAction = {
   ConfigLoader,
   Action,
   Utils,
-  helpers,
 };
 
 export default TasitAction;
