+++ packages/action/tsconfig.json
@@ -1,6 +1,9 @@
 {
   "extends": "../../tsconfig.json",
   "include": ["src", "types"],
+  "exclude": [
+    "**/*.test.ts"
+  ],
   "compilerOptions": {
     "outDir": "./dist",
     "rootDir": "./src",
