+++ packages/contract-based-account/package.json
@@ -57,5 +57,5 @@
     "tslib": "^2.0.0",
     "typescript": "^3.9.3"
   },
-  "gitHead": "fc17ed73b2c70e3f94dcc4b23ffbf502fa16b0ab"
+  "gitHead": "08af46b69dccad031c3230a563650278e8688006"
 }
