+++ packages/contracts/package.json
@@ -77,5 +77,5 @@
   "dependencies": {
     "@openzeppelin/contracts": "3.0.2"
   },
-  "gitHead": "fc17ed73b2c70e3f94dcc4b23ffbf502fa16b0ab"
+  "gitHead": "08af46b69dccad031c3230a563650278e8688006"
 }
