+++ packages/tasit/src/testHelpers/helpers.ts
@@ -1,6 +1,6 @@
 import { TasitContracts } from "../TasitSdk";
 import DecentralandUtils from "../helpers/DecentralandUtils";
-import actionHelpers from "@tasit/action/dist/testHelpers/helpers";
+import actionHelpers from "@tasit/test-helpers";
 const {
   addressesAreEqual,
   expectMinimumTokenBalances,
