+++ packages/test-helpers/test/blah.test.ts
@@ -1,0 +1,7 @@
+import { sum } from '../src';
+
+describe('blah', () => {
+  it('works', () => {
+    expect(sum(1, 1)).toEqual(2);
+  });
+});
